% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaplot_is_oa.R
\name{areaplot_is_oa}
\alias{areaplot_is_oa}
\title{areaplot number of oa/not_oa articles}
\usage{
areaplot_is_oa(data = NULL, rel_numbers = FALSE)
}
\arguments{
\item{rel_numbers}{whether absolute or relative figures should be shown}

\item{pubs_oa_year}{dataset}
}
\value{

}
\description{
areaplot number of oa/not_oa articles
}
