% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_react_table.R
\name{s2_react_table}
\alias{s2_react_table}
\title{Interactive tables for S2 summary data}
\usage{
s2_react_table(data, all = FALSE)
}
\arguments{
\item{data}{the data}
}
\description{
Interactive tables for S2 summary data
}
