% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot_oa}
\alias{scatterplot_oa}
\title{Create OA / Overall Scatterplot}
\usage{
scatterplot_oa(x, insts = NULL)
}
\arguments{
\item{x}{dataset}

\item{insts}{institutions that you want to highlight}
}
\description{
Create OA / Overall Scatterplot
}
