% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaplot_is_oa.R
\name{areaplot_is_oa}
\alias{areaplot_is_oa}
\title{areaplot number of oa/not_oa articles}
\usage{
areaplot_is_oa(data = NULL, rel_numbers = FALSE)
}
\arguments{
\item{data}{dataset}

\item{rel_numbers}{whether absolute or relative figures should be shown}
}
\description{
areaplot number of oa/not_oa articles
}
