% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot_oa}
\alias{boxplot_oa}
\title{Box plot}
\usage{
boxplot_oa(oa_shares_inst_sec_boxplot, insts = NULL)
}
\arguments{
\item{oa_shares_inst_sec_boxplot}{dataset}

\item{insts}{institutions that you want to highlight}
}
\value{

}
\description{
Box plot
}
