% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot_oa}
\alias{scatterplot_oa}
\title{Create OA / Overall Scatterplot}
\usage{
scatterplot_oa(oa_shares_inst_sector)
}
\arguments{
\item{oa_shares_inst_sector}{data}
}
\value{

}
\description{
Create OA / Overall Scatterplot
}
